Add checkpoint after DGW
I think this makes for a good checkpoint, in part because it's after the double spend attack and DGW switch, and in part because it has 1½+ minutes on either side so most likely no conflicting timestamps. Would someone mind filling out the unix timestamp and transaction count that'd be nice (Haven't bothered fixing Rackle yet, so can't check myself).